cpufreq: Make sure the cpufreq notifier is called last
The cpufreq pm notifier also sets the cpu to 800MHz (imposed by the
hardware) to suspend the device. Make sure it is called last when the
system is going to suspend, so the cpu doesn't need to spend too many
cycles at 800MHz unnecessarily.

Thanks to arighi for the original commit.